Updated to v1.0.0-rc3-1
Note: There is still an issue with the DbConnectionBuilder not containing a method called .with_token in version "1.0.0-rc3.1" of the spacetimedb-sdk. This document will be updated with the correct version when it is know. For now, pulling the latest git version is possible by using the dependancy:
spacetimedb-sdk = { git = "https://github.com/clockworklabs/SpacetimeDB" }