Small doc fixes potpourri (#198)
* Remove hedging about table access from module quickstarts

We'll announce RLS when we announce it. For now, what we have is what we have.

* Remove hedging about supported module languages

This kind of thing belongs in a roadmap, not anywhere else in our docs.

* Fix :fingers_crossed: formatting of link to scheduled reducers

* Fix link

* List module langs in alphabetical order

Which also happens to be decreasing order of support and battle-tested-ness

* Re-order various and pages in sidebar

- Internals get their own section, and move down.
- Appendix gets its own section, instead of joining "Subscriptions."
- SQL and Subscriptions move up.

* Remove outdated guidance about tokens

We don't have "SpacetimeDB tokens" anymore, we just have regular OIDC JWTs.
We don't need to offer any special guidance about JWT hygiene.